refactor logic into usePages
